requirements.txt
pydu
python-dotenv
openai
tqdm
gradio
scipy